Copied contents back from VerticalBarChart-fix.jsx to VerticalBarChart.jsx
